% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch-GUI.R
\name{ConfigureApp}
\alias{ConfigureApp}
\title{GUI for launching shiny Application for Configuration}
\usage{
ConfigureApp()
}
\description{
GUI for launching shiny Application for Configuration
}
\examples{
\dontrun{ConfigureApp()}
}
