% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VirusSimulationR.R
\docType{package}
\name{VirusSimulationR}
\alias{VirusSimulationR}
\alias{VirusSimulationR-package}
\title{VirusSimulationR - Interactive group simulation activity for viral spread in R}
\description{
Shiny application that pulls data from a Google sheet populated by student-
submitted responses from an in-class viral spread simulation activity.
Application establishes connection to Google and starting number assignments
using the configuration app, while the main app loads the data and provides
a GUI for generating visualizations.
}
